requirements.txt
APScheduler==3.10
pytz
python-binance==1.0.28
streamlit>=1.45.0
pandas
ccxt
python-telegram-bot>=20.0
SQLAlchemy>=2.0
psycopg2-binary
werkzeug>=2.0.0
ruamel.yaml>=0.17.10
PyYAML>=5.4
streamlit-authenticator>=0.2.3
itsdangerous>=2.1.2
streamlit-cookies-manager>=0.0.8
bcrypt

